pyproject.toml
[project]
name = "gemini-ai-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "reorder-python-imports>=3.15.0",
]
